Tweak source to get around warnings.

+++ unix/util/6l/asm.c
@@ -335,11 +335,11 @@ datblk(long s, long n)
 				if(debug['a'] && i == 0) {
 					Bprint(&bso, pcstr, l+s+INITDAT);
 					for(j=0; j<c; j++)
-						Bprint(&bso, "%.2ux", cast[fnuxi4[j]]);
+						Bprint(&bso, "%.2ux", cast[(int)fnuxi4[j]]);
 					Bprint(&bso, "\t%P\n", curp);
 				}
 				for(; i<c; i++) {
-					buf.dbuf[l] = cast[fnuxi4[i]];
+					buf.dbuf[l] = cast[(int)fnuxi4[i]];
 					l++;
 				}
 				break;
@@ -348,11 +348,11 @@ datblk(long s, long n)
 				if(debug['a'] && i == 0) {
 					Bprint(&bso, pcstr, l+s+INITDAT);
 					for(j=0; j<c; j++)
-						Bprint(&bso, "%.2ux", cast[fnuxi8[j]]);
+						Bprint(&bso, "%.2ux", cast[(int)fnuxi8[j]]);
 					Bprint(&bso, "\t%P\n", curp);
 				}
 				for(; i<c; i++) {
-					buf.dbuf[l] = cast[fnuxi8[i]];
+					buf.dbuf[l] = cast[(int)fnuxi8[i]];
 					l++;
 				}
 				break;
@@ -396,11 +396,11 @@ datblk(long s, long n)
 				if(debug['a'] && i == 0) {
 					Bprint(&bso, pcstr, l+s+INITDAT);
 					for(j=0; j<c; j++)
-						Bprint(&bso, "%.2ux", cast[inuxi1[j]]);
+						Bprint(&bso, "%.2ux", cast[(int)inuxi1[j]]);
 					Bprint(&bso, "\t%P\n", curp);
 				}
 				for(; i<c; i++) {
-					buf.dbuf[l] = cast[inuxi1[i]];
+					buf.dbuf[l] = cast[(int)inuxi1[i]];
 					l++;
 				}
 				break;
@@ -408,11 +408,11 @@ datblk(long s, long n)
 				if(debug['a'] && i == 0) {
 					Bprint(&bso, pcstr, l+s+INITDAT);
 					for(j=0; j<c; j++)
-						Bprint(&bso, "%.2ux", cast[inuxi2[j]]);
+						Bprint(&bso, "%.2ux", cast[(int)inuxi2[j]]);
 					Bprint(&bso, "\t%P\n", curp);
 				}
 				for(; i<c; i++) {
-					buf.dbuf[l] = cast[inuxi2[i]];
+					buf.dbuf[l] = cast[(int)inuxi2[i]];
 					l++;
 				}
 				break;
@@ -420,11 +420,11 @@ datblk(long s, long n)
 				if(debug['a'] && i == 0) {
 					Bprint(&bso, pcstr, l+s+INITDAT);
 					for(j=0; j<c; j++)
-						Bprint(&bso, "%.2ux", cast[inuxi4[j]]);
+						Bprint(&bso, "%.2ux", cast[(int)inuxi4[j]]);
 					Bprint(&bso, "\t%P\n", curp);
 				}
 				for(; i<c; i++) {
-					buf.dbuf[l] = cast[inuxi4[i]];
+					buf.dbuf[l] = cast[(int)inuxi4[i]];
 					l++;
 				}
 				break;
@@ -433,11 +433,11 @@ datblk(long s, long n)
 				if(debug['a'] && i == 0) {
 					Bprint(&bso, pcstr, l+s+INITDAT);
 					for(j=0; j<c; j++)
-						Bprint(&bso, "%.2ux", cast[inuxi8[j]]);
+						Bprint(&bso, "%.2ux", cast[(int)inuxi8[j]]);
 					Bprint(&bso, "\t%P\n", curp);
 				}
 				for(; i<c; i++) {
-					buf.dbuf[l] = cast[inuxi8[i]];
+					buf.dbuf[l] = cast[(int)inuxi8[i]];
 					l++;
 				}
 				break;
